requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
Pillow>=9.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
streamlit>=1.28.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
requests>=2.31.0
openai>=1.0.0
datasets>=2.14.0